--- style.css
@@ -1,119 +1,0 @@
-
- .box{
-  display: flex;
-  height: 900px;
-  flex-direction: row;
-  justify-content: center;
-  flex-wrap: wrap;
-  background-image: url("../../media/space.jpg");
-  background-size: contain;
-  background-repeat: repeat;
-}
-h2 
- {
-   font-size: 2vh;
-   font-weight: normal;
-   font-style: normal;
-   font-weight: 600;
-   color: rgb(243, 237, 249)
- }
-.kol1 {
- padding-top: 1vw;
- width: 25%;
- padding-left: 1vw;
- display: flex;
- background:rgba(0, 0, 0, 0.4);
- transition: .2s;
- flex-shrink: 2;
-}
-.kol2 {
- width: 50%;
- display: flex;
- flex-direction: column;
- padding-top: 2vw;
- background:rgb(23, 7, 29);
- transition: .2s;
- flex-shrink: 2;
- text-align: center;
- list-style-type: none;
-}
-.kol3 {
- padding-top: 1vw;
- width: 25%;
- padding-left: 1vw;
- display: flex;
- background:rgba(0, 0, 0, 0.4);
- transition: .2s;
- flex-shrink: 2;
-}
-.inputFields {
- margin: 10px 0;
- font-size: 1.5vh;
- padding: 2px;
- width: 20vw;
- border: 1px solid #ffd100;
- border-top: none;
- border-left: none;
- border-right: none;
- background: rgba(20, 20, 20, .2);
- color: white;
- outline: none;
-}
-#join-btn {
- border: 1px solid #45024b;
- width: 50%;
- border-radius: 20px;
- background: rgba(110, 87, 143, 0.4);
- font-size: 18px;
- font-weight: 100;
- text-align: center;
- color: rgba(243, 237, 249, 0.7);
- cursor: pointer;
- transition: .4s;
-}
-#join-btn:hover{
-   background: rgba(44, 14, 55, 0.8);
-   font-weight: 900;
-   transition: background .5s;
-}
-@media (max-width: 1450px){
- .box{
-   padding-top: 40px;
- }
-}
-@media (max-width: 1150px){
- .box{
-   padding-top: 60px;
- }
-}
-@media (max-width: 970px){
- .box{
-   padding-top: 85px;
- }
-}
-@media (max-width: 800px) {
- .kol1{
-   width: 5%;
- }
- .kol2{
-   width: 90%;
- }
- .kol3{
-   width: 5%;
- }
- .inputFields{
-   width: 80%;
-   padding: 1px;
- }
- .kol2 input{
-   font-size: 2.5vh;
- }
- .box{
-   padding-top: 100px;
- }
-}
-@media (max-width: 680px){
- .box{
-   padding-top: 125px;
- }
-}
